requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.51
langchain-ollama==0.3.1
langsmith==0.3.30
magic-filter==1.0.12
multidict==6.4.2
ollama==0.4.7
openai==1.42.0
orjson==3.10.16
packaging==24.2
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
zstandard==0.23.0

langgraph~=0.3.34
langchain-community~=0.3.22
langchain~=0.3.24